+++ src/apps/settings/modules/WindowManager/main/infra/Others.tsx
@@ -1,5 +1,6 @@
 import { SettingsGroup, SettingsOption } from '../../../../components/SettingsBox';
 import { InputNumber } from 'antd';
+import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 
 import { useAppSelector } from '../../../shared/utils/infra';
@@ -11,6 +12,7 @@ export const OthersConfigs = () => {
   const resizeDelta = useAppSelector(SeelenWmSelectors.resizeDelta);
 
   const dispatch = useDispatch();
+  const { t } = useTranslation();
 
   const onChangeResizeDelta = (value: number | null) => {
     dispatch(WManagerSettingsActions.setResizeDelta(value || 0));
@@ -20,7 +22,7 @@ export const OthersConfigs = () => {
     <>
       <SettingsGroup>
         <SettingsOption>
-          <span>Resize delta (%)</span>
+          <span>{t('wm.resize_delta')}</span>
           <InputNumber value={resizeDelta} onChange={onChangeResizeDelta} min={1} max={40} />
         </SettingsOption>
       </SettingsGroup>