+++ src/apps/settings/modules/shared/store/app/StateBridge.ts
@@ -63,6 +63,7 @@ export const StateToJsonSettings = (state: RootState): ISettings => {
     'fancyToolbar',
     'ahkVariables',
     'devTools',
+    'language',
   ]);
 };